requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
h5py>=3.0.0
pyyaml>=5.4.0
tqdm>=4.60.0
rotary-embedding-torch>=0.3.0
